LoadersFactory updated. FetchingOrderGenerator class added for ordering entities based on metadata. Main loaders was transfered to another file colled core_data_loaders.py
