+++ service_api/grabbing_api/utils/db.py
@@ -1,196 +1,119 @@
 """
-Module with data Loaders
+Module with data LoadersFactory and order generator
 """
-from abc import ABC, abstractmethod
-from typing import Any, Dict, List
-from sqlalchemy import select
-import requests
+from collections import defaultdict, deque
+from typing import Any, Dict, List, Tuple
+
 from marshmallow.exceptions import ValidationError
-from requests.exceptions import RequestException
-from service_api import session_scope
-from service_api.exceptions import ObjectNotFound, ResponseNotOk
-from service_api.grabbing_api.constants import DOMRIA_TOKEN, PATH_TO_METADATA
-from service_api.models import City, OperationType, RealtyType, State
-from service_api.schemas import (CitySchema, OperationTypeSchema,
-                                 RealtyTypeSchema, StateSchema)
+from service_api.exceptions import (CycleReferenceException, MetaDataError, ObjectNotFoundException,
+                                    ResponseNotOkException)
+from service_api.grabbing_api.constants import PATH_TO_CORE_DB_METADATA
 
-from .grabbing_utils import load_data, open_metadata
+from . import core_data_loaders
+from .grabbing_utils import open_metadata
 
 
-class BaseLoader(ABC):
+class FetchingOrderGenerator:
     """
-    Abstract class for loading static base date to DB (exmp. cities, states, realty_types...)
+    Create an sequence ready for fetching that consist of ordered entities based on entites dependencies
     """
 
-    @abstractmethod
-    def load(self, *args, **kwargs) -> None:
+    def __init__(self, dependencies: Dict[str, List]) -> None:
         """
-        Main function of retrieving and loading data to db
+        Inithializer of class
+        :params: Dict[str, List] dependencies
+            Example:
+                {
+                    "cities": ["states", "realty_type"],
+                    "states": [],
+                    "realty_type": ["states"],
+                    "operation_type": []
+                }
         """
-        return None
+        self.dependencies = dependencies
 
-
-class CityLoader(BaseLoader):
-    """
-    Loads cities to db
-    """
-
-    def load(self, *args, **kwargs) -> Dict[int, int]:
-        state_ids: List[int] = args[0]
-        with session_scope() as session:
-            smt = select(State.id).order_by(State.id)
-            state_ids = state_ids or [state_id for state_id, *_ in session.execute(smt).all()]
-        return {state_id: self.load_cities_by_state(state_id=state_id) for state_id in state_ids}
-
-    def load_cities_by_state(self, **kwargs: int) -> int:
+    def get_order(self, entities: List[str]) -> List[str]:
         """
-        Getting cities from DOMRIA by original state_id
-        Returns amount of fetched cities
-        :param: state_id: int
-        :return: int
+        Create ordered sequence based on dependencies and entities that need to be ordered
+
+        :params: List[str] entities (should be subset of dependencies.keys() from __init__)
+        :return: List[str] list of ordered entities
+
+        At first, Adjacency structure is created from all dependencies and entities.
+        During first step collection of all roots (All entities without dependencies) is filled.
+        Then directed graph is build and BFS is used for finding nedded order of entities.
         """
 
-        if (state_id := kwargs.get("state_id")) is None:
-            raise KeyError("No parameter state_id provided in function load")
+        adj, roots = defaultdict(list), []
 
-        with session_scope() as session:
-            state = session.query(State).get({"id": state_id})
+        for name in entities:
+            dependencies = sorted(set(entities) & set(self.dependencies[name]))
+            if not dependencies:
+                roots.append(name)
+                continue
+            for dependecy in dependencies:
+                if dependecy in entities:
+                    adj[dependecy].append(name)
 
-        if state is None:
-            raise ObjectNotFound(errors=["No such state_id in DB"])
+        visited = []
 
-        domria_meta = open_metadata(PATH_TO_METADATA)["DOMRIA API"]
-        domria_cities_meta = domria_meta["url_rules"]["cities"]
+        for root in roots:
+            self.bfs(adj, root, visited)
 
-        params = {
-            "lang_id": domria_meta["optional"]["lang_id"],
-            "api_key": DOMRIA_TOKEN
-        }
+        if len(visited) != len(entities):
+            raise CycleReferenceException("Cycle dependencies detected during generating entities order")
 
-        response = requests.get("{}{}/{}".format(domria_meta["base_url"],
-                                                 domria_cities_meta["url_prefix"],
-                                                 state.original_id),
-                                params=params)
-        if not response.ok:
-            raise ResponseNotOk(response.text)
+        return visited
 
-        processed_cities = []
-        for city in response.json():
-            processed_city = {key: city[external_service_key]
-                              for key, external_service_key in domria_cities_meta["filters"].items()}
-            processed_city["state_id"] = state.id
-            processed_cities.append(processed_city)
-
-        for data in processed_cities:
-            try:
-                load_data(data, City, CitySchema)
-            except ValidationError as error:
-                print(error)
-                raise
-
-        return len(processed_cities)
-
-
-class StateLoader(BaseLoader):
-    """
-    Loads sates to db
-    """
-
-    def load(self, *args, **kwargs) -> int:
+    def bfs(self, adj: Dict, node: str,  visited: List) -> List:
         """
-        Getting states from DOMRIA
-        Returns amount of fetched states
-        :return: int
+        Bypass the graph using bfs
+        :params: Dict: adj (Adjacency structure)
+                 str: start node
+                 List: list of visited nodes
         """
+        queue = deque((node,))
 
-        domria_meta = open_metadata(PATH_TO_METADATA)["DOMRIA API"]
-        domria_states_meta = domria_meta["url_rules"]["states"]
+        while queue:
+            node = queue.popleft()
+            queue.extend((next_node for next_node in adj[node] if next not in visited))
+            visited.append(node)
 
-        params = {
-            "lang_id": domria_meta["optional"]["lang_id"],
-            "api_key": DOMRIA_TOKEN
-        }
-
-        url = "{}{}".format(domria_meta["base_url"], domria_states_meta["url_prefix"])
-        response = requests.get(url, params=params)
-        if not response.ok:
-            raise RequestException(response.text)
-
-        states_json = response.json()
-        processed_states = [{key: state[external_service_key]
-                             for key, external_service_key in domria_states_meta["filters"].items()}
-                            for state in states_json]
-
-        for data in processed_states:
-            load_data(data, State, StateSchema)
-
-        return len(processed_states)
-
-
-class RealtyTypeLoader(BaseLoader):
-    """
-    Loads RealtyType from metadata
-    """
-
-    def load(self, *args, **kwargs) -> int:
-        """
-        Getting realty types from metadata
-        Returns amount of fetched realty types
-        :return: int
-        """
-        domria_meta = open_metadata(PATH_TO_METADATA)["DOMRIA API"]
-        realty_types = domria_meta["url_characteristics"]["realty_type"]
-
-        keys = domria_meta["model_characteristics"]["realty_type"]["filters"].keys()
-
-        for realty_type_name, realty_type_original_id in realty_types.items():
-            data = dict(zip(keys, (realty_type_name, realty_type_original_id)))
-            load_data(data, RealtyType, RealtyTypeSchema)
-
-        return len(realty_types)
-
-
-class OperationTypeLoader(BaseLoader):
-    """
-    Loads OperationType from metadata
-    """
-
-    def load(self, *args, **kwargs) -> int:
-        """
-        Getting operation types from metadata
-        Returns amount of fetched operation types
-        :return: int
-        """
-
-        domria_meta = open_metadata(PATH_TO_METADATA)["DOMRIA API"]
-        operation_types = domria_meta["url_characteristics"]["operation_type"]
-
-        keys = domria_meta["model_characteristics"]["operation_type"]["filters"].keys()
-
-        for operation_type_name, operation_type_original_id in operation_types.items():
-            data = dict(zip(keys, (operation_type_name, operation_type_original_id)))
-            load_data(data, OperationType, OperationTypeSchema)
-
-        return len(operation_types)
+        return visited
 
 
 class LoadersFactory:
     """
     Load core data to db based on str input
     """
-    __mapper: Dict[str, BaseLoader] = {
-        "cities": CityLoader,
-        "states": StateLoader,
-        "realty_types": RealtyTypeLoader,
-        "operation_types": OperationTypeLoader
-    }
+
+    def __init__(self) -> None:
+        """
+        Initialization of LoadersFactory thst uses metadata for configuration
+        This class can raise MetaDataError
+        """
+        metadata = open_metadata(PATH_TO_CORE_DB_METADATA)["CORE DATA"]
+
+        for info in metadata.values():
+            info["loader"] = getattr(core_data_loaders, info.get("loader", None), None)
+
+        self.__mapper = metadata
 
     def get_available_entites(self) -> List[str]:
         """
-        Returns all possible name of entities stored in mapper
+        Returns all possible name of entities that can be loaded
         :return: list[str] with name of entities
         """
         return self.__mapper.keys()
+
+    def divide_input_entities(self, entities: List[str]) -> Tuple[List, List]:
+        """
+        This function divides given envities on 2 groups (entities that can be loaded, unknown entities)
+        :params: List[str]
+        :return: tuple[List, List]
+        """
+        all_entities, input_entities = set(self.__mapper), set(entities)
+        return sorted(all_entities & input_entities), sorted(input_entities - all_entities)
 
     def load(self, **entities_to_load: List) -> Dict[str, Any]:
         """
@@ -201,13 +124,23 @@ class LoadersFactory:
                                 Any - json-like status info about fetching data to db
         """
 
-        statuses = {}
-        for entity, values in entities_to_load.items():
+        can_be_loaded, unknown = self.divide_input_entities(entities_to_load)
+
+        try:
+            ordered_entities = FetchingOrderGenerator(
+                {key: info["depends_on"] for key, info in self.__mapper.items()}).get_order(can_be_loaded)
+        except CycleReferenceException as error:
+            print(error.args, error.desc)
+            raise MetaDataError(desc=f"Metadata that is used: {PATH_TO_CORE_DB_METADATA}") from error
+
+        statuses = {key: {"status": "Unknown entity"} for key in unknown}
+        for entity in ordered_entities:
             try:
-                statuses[entity] = {"status": "SUCCESSFUL", "data": self.__mapper[entity]().load(values)}
-            except ObjectNotFound as error:
-                error_mesaage = error.errors
-            except ResponseNotOk as error:
+                statuses[entity] = {"status": "SUCCESSFUL",
+                                    "data": self.__mapper[entity]["loader"]().load(entities_to_load[entity])}
+            except ObjectNotFoundException as error:
+                error_mesaage = error.args
+            except ResponseNotOkException as error:
                 error_mesaage = error.args
             except KeyError as error:
                 error_mesaage = error.args
